+++ packages/kuflow-test/test/tasks.test.ts
@@ -29,7 +29,7 @@ import {
   TaskDeleteElementValueDocumentCommand,
   TaskSaveElementCommand,
   TaskSaveElementValueDocumentCommand,
-} from '@kuflow/kuflow-rest-client'
+} from '@kuflow/kuflow-rest'
 import { randomUUID } from 'crypto'
 import nock from 'nock'