+++ packages/kuflow-temporal-core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @kuflow/kuflow-engine-client-core
+# @kuflow/kuflow-temporal-core
 
 ## 0.2.0-next.3
 
@@ -7,7 +7,7 @@
 - 5f613df: Integrate kuflow-rest-client
 - 2cecc38: Update to the latest version of kuflow-rest-client
 - Updated dependencies [5f613df]
-  - @kuflow/kuflow-rest-client@0.2.0-next.3
+  - @kuflow/kuflow-rest@0.2.0-next.3
 
 ## 0.2.0-next.2