+++ packages/kuflow-test/CHANGELOG.md
@@ -6,6 +6,6 @@
 
 - Updated dependencies [5f613df]
 - Updated dependencies [2cecc38]
-  - @kuflow/kuflow-engine-client-activity-kuflow@0.2.0-next.3
-  - @kuflow/kuflow-engine-client-core@0.2.0-next.3
-  - @kuflow/kuflow-rest-client@0.2.0-next.3
+  - @kuflow/kuflow-temporal-activity-kuflow@0.2.0-next.3
+  - @kuflow/kuflow-temporal-core@0.2.0-next.3
+  - @kuflow/kuflow-rest@0.2.0-next.3